qfile-dom0-unpacker: support waiting for free space
Add support for -w option, which makes the extract process to wait for
enough free space in the target directory. This is especially useful
when restoring a backup, which is multi-stage asynchronous process -
wait for another process to process extracted files (and free the space)
before extracting more.

QubesOS/qubes-issues#5478