Better parsing of command-line options
This makes qubes-dom0-update aware of various command-line options that
take arguments, so it can process them accordingly.  It is not a perfect
solution; that would require using a C wrapper program to reject
abbreviated or unknown options.  It also uses bash regex matching
instead of the `echo | grep` anti-pattern, replaces an `==` with `=`,
adds a missing `--`, and unsets a variable that will be checked for
being set later.

The same approach used here could be used to refactor
qubes-gpg-client-wrapper if it were worth the time, which it probably
isn’t unless qubes-gpg-client-wrapper will see future use.