vmupdate: make log group-writable
When qubes-vm-update is called as root, the log would be created as
root, breaking the usage by normal user. Fix this by trying to make the
log writable by the "qubes" group. This may fail if the tool is called
as non-root user, but that's okay - at this point the current
permissions were enough to start the tool.