vmupdate: Apply early fix for meminfo-writer SELinux label
The relevant package updates are supposed to fix the label anyway, but
before that the template has too little memory to apply the update
reliably (most of the time it works, but not always). Apply the label
fix early to have more memory when installing updates.

Script by Minimalist <suspectindustrie@protonmail.com>

QubesOS/qubes-issues#9663