vmupdate: Fix SELinux plugin detection
The current plugin always run if the system is detected as fedora.
But, SELinux is not always installed (minimal templates for example)
or it can be intentionally disabled by the user, meaning the script will throw a chcon error in the update log.
Fix that by first checking if SELinux is installed, and then check if the status is Enforcing/Permissive.