vmupdate: Send SIGUSR1 to meminfo-writer
The current implementation restart the meminfo-writer service after
changing the selinux label which can sometime take a while for the
memory to go back to normal.
Instead, send a USR1 signal to the meminfo-writer process to change the
memory immediately, which also assure smooth selinux updates
installation in the future.